+++ config/database.json
@@ -1,1 +1,1 @@
-[]
+[{"id":"ffbcac75-9f72-48c1-b02a-4ec1435dc0e5","name":"Default 1","description":"this is default 1 ","imageUrl":"https://upload.wikimedia.org/wikipedia/en/a/af/Star_Wars_The_Rise_of_Skywalker_poster.jpg","difficulty":"1"},{"id":"df2455b2-5fb5-43dd-b3b6-97b5c6789697","name":"Default 2","description":"this is default 2","imageUrl":"https://upload.wikimedia.org/wikipedia/en/a/af/Star_Wars_The_Rise_of_Skywalker_poster.jpg","difficulty":"2"},{"id":"e7130023-e625-4cdb-a837-a76d6f99b762","name":"Default 3","description":"this is Default 3","imageUrl":"https://upload.wikimedia.org/wikipedia/en/a/af/Star_Wars_The_Rise_of_Skywalker_poster.jpg","difficulty":"3"}]
