+++ views/index.hbs
@@ -7,14 +7,22 @@
     <input type="number" name="to" class="difficulty" placeholder="Difficulty to...">
     <input type="submit" value="search">
   </form>
-  <div class="cube">
+  {{#each cubes}}
+  {{> cube}}
+
+  {{else}}
+    <p>No Cubes</p>
+  {{/each}}
+  {{!-- <div class="cube">
     <img class="cube"
       src="https://ae01.alicdn.com/kf/HTB1CSddXRxRMKJjy0Fdq6yifFXa6/Gan-356-Air-SM-3x3-Black-Magic-cube-GAN-Air-SM-Magnetic-3x3x3-Speed-cube-gans.jpg">
     <p class="name">Gan356 Air SM</p>
     <p><span>Difficulty level:</span> 3</p>
     <a class="btn" href="/details/5c39f1851ab3b24f5c03e182">Details</a>
-  </div>
-  <div class="cube">
+  </div> --}}
+
+
+  {{!-- <div class="cube">
     <img class="cube" src="https://thingsidesire.com/wp-content/uploads/2018/06/Eco-Dark-Rubik%E2%80%99s-Cube2.jpg">
     <p class="name">Eco-Dark</p>
     <p><span>Difficulty level:</span> 6</p>
@@ -31,5 +39,5 @@
     <p class="name">Megaminx</p>
     <p><span>Difficulty level:</span> 3</p>
     <a class="btn" href="/details/5c3a68d9f78d6638ec8fdad7">Details</a>
-  </div>
+  </div> --}}
 </main>
