+++ views/partials/cube.hbs
@@ -1,0 +1,6 @@
+<div class="cube">
+    <img class="cube" src="{{imageUrl}}">
+    <p class="name">{{name}}</p>
+    <p><span>Difficulty level:</span>{{difficulty}}</p>
+    <a class="btn" href="/details/{{id}}">Details</a>
+</div>
